loans pause and reject

+++ hrms-project/src/app/leave-master/leave.service.ts
@@ -788,7 +788,13 @@ generateAttendanceReport(schema: string, data: any): Observable<any> {
 
 
   deletePayroll(payrollId: number): Observable<any> {
-    const url = `${this.apiUrl}/payroll/api/PayrollRun/${payrollId}/`;
+    const selectedSchema = localStorage.getItem('selectedSchema');
+    if (!selectedSchema) {
+      console.error('No schema selected.');
+      return throwError('No schema selected.'); // Return an error observable if no schema is selected
+    }
+
+    const url = `${this.apiUrl}/payroll/api/PayrollRun/${payrollId}/?schema=${selectedSchema}`;
     return this.http.delete(url);
   }
 }
